Address review fixes for resilience regen (#731)
Centralize semver build metadata stripping and add regression coverage for legacy inline annotation extraction.
